+++ apps/decodex/src/orchestrator/lane_decision.rs
@@ -1,6 +1,9 @@
+mod decision_model;
 mod json_projection;
 mod model;
 mod projection;
+mod snapshot_json;
+mod snapshot_observation;
 
 pub(super) use self::{
 	model::{LaneDecisionSnapshot, LaneNextAction},
