+++ apps/decodex/src/orchestrator/lane_decision/model.rs
@@ -1,15 +1,6 @@
-use serde_json::Value;
-
 use crate::orchestrator::{
 	IssueDispatchMode, PhaseGoalKind, RepoGateFailureDisposition, RetryKind,
-	kernel::{
-		action::OwnedLaneAction,
-		command::CommandIntentKind,
-		decision::OwnedLaneDecision,
-		facts::LaneObservation,
-		state::{LivenessState, TerminalizationState},
-	},
-	lane_decision::{json_projection, projection},
+	kernel::decision::OwnedLaneDecision,
 };
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
@@ -148,61 +139,6 @@ impl LaneDecisionSnapshot {
 			terminal_evidence_present: false,
 		}
 	}
-
-	pub(in crate::orchestrator) fn to_json(&self, action: LaneNextAction, reason: &str) -> Value {
-		let decision = projection::decide_lane_next_action(self);
-
-		serde_json::json!({
-			"schema": "decodex.lane_decision_snapshot/1",
-			"issue_identifier": self.issue_identifier,
-			"run_id": self.run_id,
-			"attempt_number": self.attempt_number,
-			"dispatch_mode": self.dispatch_mode.as_str(),
-			"active_phase": self.active_phase.map(PhaseGoalKind::as_str),
-			"continuation_pending": self.continuation_pending,
-			"retry_kind": self.retry_kind.map(RetryKind::as_str),
-			"retry_budget_consumed": self.retry_budget_consumed,
-			"progress_blocker_count": self.progress_blocker_count,
-			"non_goal_violation": self.non_goal_violation,
-			"phase_acceptance_failure": self.phase_acceptance_failure,
-			"repo_gate_disposition": self.repo_gate_disposition.map(RepoGateFailureDisposition::as_str),
-			"scope_envelope_violation": self.scope_envelope_violation,
-			"ambiguous_lineage": self.ambiguous_lineage,
-			"terminal_evidence_present": self.terminal_evidence_present,
-			"next_action": action.as_str(),
-			"reason": reason,
-			"kernel_decision": json_projection::owned_lane_decision_to_json(&decision.kernel_decision),
-		})
-	}
-
-	pub(super) fn to_kernel_observation(&self) -> LaneObservation {
-		let mut observation = LaneObservation::for_issue(self.issue_identifier.clone());
-
-		observation.run_id = Some(self.run_id.clone());
-		observation.authority_complete = true;
-		observation.run_lease = true;
-		observation.active_owned_work = true;
-		observation.liveness = LivenessState::ThreadActive;
-		observation.terminalization = if self.terminal_evidence_present {
-			TerminalizationState::CleanupPending
-		} else {
-			TerminalizationState::None
-		};
-		observation.contradictory_authority = self.ambiguous_lineage;
-		observation.human_attention_signal = self.progress_blocker_count > 0
-			|| self.non_goal_violation
-			|| self.scope_envelope_violation
-			|| self.repo_gate_disposition == Some(RepoGateFailureDisposition::NeedsHumanAttention);
-		observation.retry_budget_available = self.phase_acceptance_failure
-			|| self.retry_kind.is_some()
-			|| self.repo_gate_disposition == Some(RepoGateFailureDisposition::ContinueRepair);
-		observation.retry_budget_exhausted = false;
-		observation.retained_lane_reusable = self.continuation_pending;
-		observation.external_signal_pending =
-			self.repo_gate_disposition == Some(RepoGateFailureDisposition::RetryAfterBackoff);
-
-		observation
-	}
 }
 
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -218,26 +154,5 @@ impl LaneDecision {
 		kernel_decision: OwnedLaneDecision,
 	) -> Self {
 		Self { next_action, reason, kernel_decision }
-	}
-
-	pub(in crate::orchestrator) fn permits_child_exit_retry_kind(&self, kind: RetryKind) -> bool {
-		let required_intent = match kind {
-			RetryKind::Continuation => CommandIntentKind::ResumeRetainedLane,
-			RetryKind::Failure => CommandIntentKind::ScheduleRetry,
-		};
-
-		self.has_command_intent(required_intent)
-	}
-
-	pub(in crate::orchestrator) fn permits_phase_repair_retry(&self) -> bool {
-		self.has_command_intent(CommandIntentKind::ScheduleRetry)
-	}
-
-	pub(in crate::orchestrator) fn blocks_automatic_execution(&self) -> bool {
-		self.kernel_decision.decision_class == OwnedLaneAction::ManualInterventionRequired
-	}
-
-	fn has_command_intent(&self, kind: CommandIntentKind) -> bool {
-		self.kernel_decision.command_intents.iter().any(|intent| intent.kind == kind)
 	}
 }
